+++ .gitignore
@@ -1,0 +1,4 @@
+pkg/
+api.key
+.api.key.trimmed
+*-handin.tar.gz
