+++ src/go.mod
@@ -1,3 +1,3 @@
 module 6.824
 
-go 1.20
+go 1.15
