+++ src/main/mrcoordinator.go
@@ -9,23 +9,21 @@ package main
 // Please do not change this file.
 //
 
-import (
-	"fmt"
-	"os"
-	"time"
-
-	"6.824/mr"
-)
+import "6.824/mr"
+import "time"
+import "os"
+import "fmt"
 
 func main() {
 	if len(os.Args) < 2 {
 		fmt.Fprintf(os.Stderr, "Usage: mrcoordinator inputfiles...\n")
 		os.Exit(1)
 	}
-	//fmt.Print(os.Args, len(os.Args))
+
 	m := mr.MakeCoordinator(os.Args[1:], 10)
 	for m.Done() == false {
 		time.Sleep(time.Second)
 	}
+
 	time.Sleep(time.Second)
 }
