+++ src/main/mrworker.go
@@ -10,28 +10,27 @@ package main
 // Please do not change this file.
 //
 
-import (
-	"fmt"
-	"log"
-	"os"
-	"plugin"
-
-	"6.824/mr"
-)
+import "6.824/mr"
+import "plugin"
+import "os"
+import "fmt"
+import "log"
 
 func main() {
-	if len(os.Args) < 2 {
+	if len(os.Args) != 2 {
 		fmt.Fprintf(os.Stderr, "Usage: mrworker xxx.so\n")
 		os.Exit(1)
 	}
-	//fmt.Print(os.Args[1])
+
 	mapf, reducef := loadPlugin(os.Args[1])
 
 	mr.Worker(mapf, reducef)
 }
 
+//
 // load the application Map and Reduce functions
 // from a plugin file, e.g. ../mrapps/wc.so
+//
 func loadPlugin(filename string) (func(string, string) []mr.KeyValue, func(string, []string) string) {
 	p, err := plugin.Open(filename)
 	if err != nil {
