+++ src/models/kv.go
@@ -1,11 +1,8 @@
 package models
 
-import (
-	"fmt"
-	"sort"
-
-	"6.824/porcupine"
-)
+import "6.824/porcupine"
+import "fmt"
+import "sort"
 
 type KvInput struct {
 	Op    uint8 // 0 => get, 1 => put, 2 => append
