+++ src/mrapps/wc.go
@@ -6,19 +6,18 @@ package main
 // go build -buildmode=plugin wc.go
 //
 
-import (
-	"strconv"
-	"strings"
-	"unicode"
+import "6.824/mr"
+import "unicode"
+import "strings"
+import "strconv"
 
-	"6.824/mr"
-)
-
+//
 // The map function is called once for each file of input. The first
 // argument is the name of the input file, and the second is the
 // file's complete contents. You should ignore the input file name,
 // and look only at the contents argument. The return value is a slice
 // of key/value pairs.
+//
 func Map(filename string, contents string) []mr.KeyValue {
 	// function to detect word separators.
 	ff := func(r rune) bool { return !unicode.IsLetter(r) }
@@ -34,9 +33,11 @@ func Map(filename string, contents string) []mr.KeyValue {
 	return kva
 }
 
+//
 // The reduce function is called once for each key generated by the
 // map tasks, with a list of all the values created for that key by
 // any map task.
+//
 func Reduce(key string, values []string) string {
 	// return the number of occurrences of this word.
 	return strconv.Itoa(len(values))
