+++ src/raft/test_test.go
@@ -8,14 +8,12 @@ package raft
 // test with the original before submitting.
 //
 
-import (
-	"fmt"
-	"math/rand"
-	"sync"
-	"sync/atomic"
-	"testing"
-	"time"
-)
+import "testing"
+import "fmt"
+import "time"
+import "math/rand"
+import "sync/atomic"
+import "sync"
 
 // The tester generously allows solutions to complete elections in one second
 // (much more than the paper's range of timeouts).
@@ -148,8 +146,10 @@ func TestBasicAgree2B(t *testing.T) {
 	cfg.end()
 }
 
+//
 // check, based on counting bytes of RPCs, that
 // each command is sent to each peer just once.
+//
 func TestRPCBytes2B(t *testing.T) {
 	servers := 3
 	cfg := make_config(t, servers, false, false)
@@ -181,7 +181,9 @@ func TestRPCBytes2B(t *testing.T) {
 	cfg.end()
 }
 
+//
 // test just failure of followers.
+//
 func For2023TestFollowerFailure2B(t *testing.T) {
 	servers := 3
 	cfg := make_config(t, servers, false, false)
@@ -226,7 +228,9 @@ func For2023TestFollowerFailure2B(t *testing.T) {
 	cfg.end()
 }
 
+//
 // test just failure of leaders.
+//
 func For2023TestLeaderFailure2B(t *testing.T) {
 	servers := 3
 	cfg := make_config(t, servers, false, false)
@@ -266,8 +270,10 @@ func For2023TestLeaderFailure2B(t *testing.T) {
 	cfg.end()
 }
 
+//
 // test that a follower participates after
 // disconnect and re-connect.
+//
 func TestFailAgree2B(t *testing.T) {
 	servers := 3
 	cfg := make_config(t, servers, false, false)
@@ -796,6 +802,7 @@ func TestPersist32C(t *testing.T) {
 	cfg.end()
 }
 
+//
 // Test the scenarios described in Figure 8 of the extended Raft paper. Each
 // iteration asks a leader, if there is one, to insert a command in the Raft
 // log.  If there is a leader, that leader will fail quickly with a high
@@ -804,6 +811,7 @@ func TestPersist32C(t *testing.T) {
 // alive servers isn't enough to form a majority, perhaps start a new server.
 // The leader in a new term may try to finish replicating log entries that
 // haven't been committed yet.
+//
 func TestFigure82C(t *testing.T) {
 	servers := 5
 	cfg := make_config(t, servers, false, false)
@@ -1184,9 +1192,11 @@ func TestSnapshotInstallUnCrash2D(t *testing.T) {
 	snapcommon(t, "Test (2D): install snapshots (unreliable+crash)", false, false, true)
 }
 
+//
 // do the servers persist the snapshots, and
 // restart using snapshot along with the
 // tail of the log?
+//
 func TestSnapshotAllCrash2D(t *testing.T) {
 	servers := 3
 	iters := 5
